refactor(ir): use unsigned type for allocation sizes (#397)
* refactor(ir): use unsigned type for allocation sizes (#394)

Add core.u64 type and use it for allocation sizes throughout the
native backend pipeline. Allocation sizes are inherently non-negative,
so u64 is semantically more accurate than i64.

- Add core.u64 dialect type with Cranelift mapping to I64
- Change RC_HEADER_SIZE from i64 to u64
- Change tribute_rt.release alloc_size attr from i64 to u64
- Update runtime FFI: __tribute_alloc/__tribute_dealloc take u64
- Update all native passes to emit core.u64 for size constants


* refactor(ir): remove unnecessary core.u64 type

TrunkIR's I<N> types are signless integers (like MLIR), so a
separate U64 type is redundant. Revert the core.u64 addition while
keeping the Rust-level u64 changes (RC_HEADER_SIZE, alloc_size attr,
runtime FFI signatures).


---------

+++ crates/tribute-passes/src/native/adt_rc_header.rs
@@ -103,8 +103,8 @@ impl<'db> RewritePattern<'db> for StructNewPattern<'db> {
         let mut ops: Vec<Operation<'db>> = Vec::new();
 
         // 1. Compute allocation size (payload + RC header)
-        let alloc_size = layout.total_size as i64 + RC_HEADER_SIZE;
-        let size_op = clif::iconst(db, location, i64_ty, alloc_size);
+        let alloc_size = layout.total_size as u64 + RC_HEADER_SIZE;
+        let size_op = clif::iconst(db, location, i64_ty, alloc_size as i64);
         let size_val = size_op.result(db);
         ops.push(size_op.as_operation());
 
@@ -127,7 +127,7 @@ impl<'db> RewritePattern<'db> for StructNewPattern<'db> {
         ops.push(store_rtti.as_operation());
 
         // 5. Compute payload pointer = raw_ptr + 8
-        let hdr_size = clif::iconst(db, location, i64_ty, RC_HEADER_SIZE);
+        let hdr_size = clif::iconst(db, location, i64_ty, RC_HEADER_SIZE as i64);
         ops.push(hdr_size.as_operation());
         let payload_ptr = clif::iadd(db, location, raw_ptr, hdr_size.result(db), ptr_ty);
         let payload_val = payload_ptr.result(db);
