+++ crates/tribute-passes/src/native/rc_lowering.rs
@@ -194,7 +194,7 @@ fn lower_nested_regions<'db>(
 struct ReleaseInfo<'db> {
     is_zero_val: Value<'db>,
     payload_ptr_val: Value<'db>,
-    alloc_size: i64,
+    alloc_size: u64,
     free_block_id: BlockId,
 }
 
@@ -417,7 +417,7 @@ fn gen_retain_inline<'db>(
     let mut ops = Vec::new();
 
     // rc_addr = ptr - RC_HEADER_SIZE
-    let hdr_sz = clif::iconst(db, location, i64_ty, RC_HEADER_SIZE);
+    let hdr_sz = clif::iconst(db, location, i64_ty, RC_HEADER_SIZE as i64);
     ops.push(hdr_sz.as_operation());
     let rc_addr = clif::isub(db, location, ptr, hdr_sz.result(db), ptr_ty);
     ops.push(rc_addr.as_operation());
@@ -463,7 +463,7 @@ fn gen_release_decrement<'db>(
     let mut ops = Vec::new();
 
     // rc_addr = ptr - RC_HEADER_SIZE
-    let hdr_sz = clif::iconst(db, location, i64_ty, RC_HEADER_SIZE);
+    let hdr_sz = clif::iconst(db, location, i64_ty, RC_HEADER_SIZE as i64);
     ops.push(hdr_sz.as_operation());
     let rc_addr = clif::isub(db, location, ptr, hdr_sz.result(db), ptr_ty);
     ops.push(rc_addr.as_operation());
@@ -506,7 +506,7 @@ fn gen_deep_release_call<'db>(
     db: &'db dyn salsa::Database,
     location: trunk_ir::Location<'db>,
     payload_ptr: Value<'db>,
-    alloc_size: i64,
+    alloc_size: u64,
     continue_block: Block<'db>,
 ) -> Vec<Operation<'db>> {
     let i64_ty = core::I64::new(db).as_type();
@@ -515,7 +515,7 @@ fn gen_deep_release_call<'db>(
     let mut ops = Vec::new();
 
     // size = iconst(alloc_size)
-    let size = clif::iconst(db, location, i64_ty, alloc_size);
+    let size = clif::iconst(db, location, i64_ty, alloc_size as i64);
     ops.push(size.as_operation());
 
     // call @__tribute_deep_release(payload_ptr, size)
